+++ packages/components/charts/src/area-gradient.tsx
@@ -1,0 +1,28 @@
+export type AreaGradientProps = {
+  id?: string
+  color?: string
+  withGradient?: boolean
+  fillOpacity: number
+}
+
+export const AreaGradient = ({
+  color,
+  id,
+  withGradient,
+  fillOpacity,
+}: AreaGradientProps) => {
+  return (
+    <>
+      {withGradient ? (
+        <linearGradient id={id} x1="0" y1="0" x2="0" y2="1">
+          <stop offset="0%" stopColor={color} stopOpacity={fillOpacity} />
+          <stop offset="100%" stopColor={color} stopOpacity={0.01} />
+        </linearGradient>
+      ) : (
+        <linearGradient id={id} x1="0" y1="0" x2="0" y2="1">
+          <stop stopColor={color} stopOpacity={fillOpacity} />
+        </linearGradient>
+      )}
+    </>
+  )
+}