+++ packages/core/src/providers/theme-provider.tsx
@@ -53,6 +53,11 @@ type ThemeProviderOptions = {
    * @default 'themeSchemeManager.localStorage'
    */
   themeSchemeManager?: ThemeSchemeManager
+  /**
+   * Key of value saved in storage.
+   * By default, it is saved to `local storage`.
+   */
+  storageKey?: string
 }
 
 export type ThemeProviderProps = Omit<EmotionThemeProviderProps, "theme"> &
@@ -62,10 +67,11 @@ export const ThemeProvider: FC<ThemeProviderProps> = ({
   theme: initialTheme = {},
   config,
   themeSchemeManager = localStorage,
+  storageKey,
   children,
 }) => {
   const [themeScheme, setThemeScheme] = useState<UITheme["themeSchemes"]>(
-    themeSchemeManager.get(config?.initialThemeScheme),
+    themeSchemeManager.get(config?.initialThemeScheme)(storageKey),
   )
 
   const changeThemeScheme: ChangeThemeScheme = useCallback(
@@ -80,9 +86,9 @@ export const ThemeProvider: FC<ThemeProviderProps> = ({
 
       setThemeScheme(themeScheme)
 
-      themeSchemeManager.set(themeScheme)
+      themeSchemeManager.set(themeScheme)(storageKey)
     },
-    [config, themeSchemeManager],
+    [config, themeSchemeManager, storageKey],
   )
 
   const theme = useMemo(
@@ -91,10 +97,10 @@ export const ThemeProvider: FC<ThemeProviderProps> = ({
   )
 
   useEffect(() => {
-    const managerValue = themeSchemeManager.get()
+    const managerValue = themeSchemeManager.get()(storageKey)
 
     if (managerValue) changeThemeScheme(managerValue)
-  }, [changeThemeScheme, themeSchemeManager])
+  }, [changeThemeScheme, themeSchemeManager, storageKey])
 
   return (
     <EmotionThemeProvider theme={{ themeScheme, changeThemeScheme, ...theme }}>