+++ packages/theme/src/components/area-chart.ts
@@ -2,6 +2,10 @@ import type { ComponentMultiStyle } from "@yamada-ui/core"
 
 export const AreaChart: ComponentMultiStyle = {
   baseStyle: {
+    container: {
+      w: "4xl",
+      h: "sm",
+    },
     grid: {},
     xAxis: {},
     yAxis: {},