+++ packages/core/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @yamada-ui/core
 
+## 1.3.2
+
+### Patch Changes
+
+- Updated dependencies [[`d8febe5`](https://github.com/hirotomoyamada/yamada-ui/commit/d8febe550ef6159cdb1ddbcbaff1a56e310a6529)]:
+  - @yamada-ui/utils@1.0.5
+  - @yamada-ui/portal@1.0.5
+
+## 1.3.1
+
+### Patch Changes
+
+- [#698](https://github.com/hirotomoyamada/yamada-ui/pull/698) [`398124a`](https://github.com/hirotomoyamada/yamada-ui/commit/398124ac60b0481c92c9ab983d932dd405c9d7ab) Thanks [@hirotomoyamada](https://github.com/hirotomoyamada)! - Added variant `unstyled` to `Resizable` component.
+
+- [#697](https://github.com/hirotomoyamada/yamada-ui/pull/697) [`242afe6`](https://github.com/hirotomoyamada/yamada-ui/commit/242afe6508f50d67780f36b1b411810cb08a7238) Thanks [@hirotomoyamada](https://github.com/hirotomoyamada)! - Fixed a bug where `storageKey` can not be changed.
+
 ## 1.3.0
 
 ### Minor Changes